Fix a bug related to attributes on players.
This mixin is used when running under Fabric to add a custom attribute
to players to tweak the reach distance.  However, it isn't needed under
forge and causes an issue where other mods' custom attributes don't get
added to new players properly.  I just added a check so that it doesn't
run under Forge which seems to fix the issue, at least with Botania.